Refactor, import/export Firebase functions
