+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "bike-kollective",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "start": "expo start --dev-client",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
     "eject": "expo eject",
     "build": "expo build:web",
@@ -21,11 +20,14 @@
     ]
   },
   "dependencies": {
+    "@react-native-google-signin/google-signin": "^8.0.0",
     "@react-navigation/native": "^6.0.10",
     "@react-navigation/native-stack": "^6.6.2",
     "expo": "~45.0.0",
+    "expo-image-picker": "~13.1.1",
+    "expo-splash-screen": "~0.15.1",
     "expo-status-bar": "~1.3.0",
-    "firebase": "^9.8.4",
+    "firebase": "^9.9.0",
     "http-module": "^0.0.5",
     "react": "17.0.2",
     "react-dom": "17.0.2",
