+++ src/client/screens/HomeScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text, View } from 'react-native';
+import { View } from 'react-native';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { RootStackParamList } from '../../../App';
 
@@ -7,11 +7,7 @@ type HomeScreenProps = NativeStackScreenProps<RootStackParamList, 'HomeScreen'>;
 
 // can't have unused parameter. Add "props" later
 const HomeScreen: React.FC<HomeScreenProps> = () => {
-  return (
-    <View>
-      <Text>The home screen</Text>
-    </View>
-  );
+  return <View></View>;
 };
 
 export default HomeScreen;
